IS-1605: Fix tests + refactor
