+++ src/test/kotlin/no/nav/syfo/testhelper/ExternalMockEnvironment.kt
@@ -1,8 +1,11 @@
 package no.nav.syfo.testhelper
 
 import no.nav.syfo.application.ApplicationState
+import no.nav.syfo.client.azuread.AzureAdClient
+import no.nav.syfo.client.pdfgen.PdfGenClient
+import no.nav.syfo.client.pdl.PdlClient
 import no.nav.syfo.client.wellknown.WellKnown
-import no.nav.syfo.testhelper.mock.*
+import no.nav.syfo.testhelper.mock.mockHttpClient
 import java.nio.file.Paths
 
 fun wellKnownInternalAzureAD(): WellKnown {
@@ -20,6 +23,19 @@ class ExternalMockEnvironment private constructor() {
     val environment = testEnvironment()
     val mockHttpClient = mockHttpClient(environment = environment)
     val wellKnownInternalAzureAD = wellKnownInternalAzureAD()
+    val azureAdClient = AzureAdClient(
+        azureEnvironment = environment.azure,
+        httpClient = mockHttpClient,
+    )
+    val pdfgenClient = PdfGenClient(
+        pdfGenBaseUrl = environment.clients.isaktivitetskravpdfgen.baseUrl,
+        httpClient = mockHttpClient,
+    )
+    val pdlClient = PdlClient(
+        azureAdClient = azureAdClient,
+        pdlEnvironment = environment.clients.pdl,
+        httpClient = mockHttpClient,
+    )
 
     companion object {
         val instance: ExternalMockEnvironment = ExternalMockEnvironment()
